Dynamic compilation bug fix
Symptom: One the first modification to a source file starts a dynamic
compilation. After that, changes are ignored.

Fix: The problem was that we need to override the "FileHash" value returned
for csproj files to include the hash of all source files. This is so that
any changes to any source file will notify ASP.NET that the .csproj file
needs to be recompiled using a the corresponding build provider.

--HG--
branch : dev