Tweak IAssemblyLoader loadig policy
A few common .net framework assemblies are referenced by the Orchard.Framework
assembly. Look into those to see if we can find the assembly we are looking for.

This is useful for orchard modules project files, where commonly used assemblies
don't need to be specified with a strong name.

--HG--
branch : dev