Feature list not refreshed when adding a new module
Update cache manager to rely on a thread static variable to
keep track of nested calls to ICacheManager.Get().

This fixes this particular bug of IExtensionManager caching the list
of available features: instead of storing the result of the computation
in a field (which is incorrect), use the cache manager.

Work Items: 16867

--HG--
branch : dev