Ensure Azure specific IStorageProvider is registered properly
The AzureBlobStorageProvider components must be registered as a
per-lifetime-scope in the "Shell" container, not in the "Host" container.

This is why it must be registered in the "Config/Sites.Config" file instead
of the "Config/Host.config" files.

This also allows us to remove the "AZURE" compilation directive from
Orchard.Framework implementation.

All in all, this ensures that the files stored in the "Media" folder
are stored in the Azure Blob Storage even when running locally in
Visual Studio.

--HG--
branch : dev
extra : transplant_source : %3E%E2%82g%18M%AB%26xi%CA%19%99H%3A%A0%20%F0%B2F