#17923: Modules from ~/Bin and ~/Core have higher priority
Because of the way ASP.NET BuildManager works(*), we can't enable
dynamic compilation of a module if the module is already
present in ~/bin.

(*) When invoking the compiler, BuildManager automatically adds
references to all assemblies present in ~/bin. So, if you have module
"X" in "~/bin", there is no way to enable dynamic compilation of
module "X" by compiling "X.csproj", as there will be tons of
compilation errors with equivalent types defined in multiple
assemblies.

--HG--
branch : 1.x